🐛 Moved Config params to Shared
fixes #47
