+++ PlugHub.Shared/Models/Configuration/UserConfigServiceParams.cs
@@ -1,0 +1,9 @@
+﻿using PlugHub.Shared.Interfaces.Services;
+using PlugHub.Shared.Models;
+using System.Text.Json;
+
+namespace PlugHub.Shared.Models.Configuration
+{
+    public record UserConfigServiceParams(string? UserConfigUriOverride = null, string? ConfigUriOverride = null, Token? Owner = null, Token? Read = null, Token? Write = null, JsonSerializerOptions? JsonSerializerOptions = null, bool ReloadOnChange = false)
+        : FileConfigServiceParams(ConfigUriOverride, Owner, Read, Write, JsonSerializerOptions, ReloadOnChange), IConfigServiceParams;
+}
