+++ PlugHub.UnitTests/Accessors/Configuration/SecureFileConfigAccessorTests.cs
@@ -9,6 +9,7 @@ using PlugHub.Shared.Interfaces.Models;
 using PlugHub.Shared.Interfaces.Platform.Storage;
 using PlugHub.Shared.Interfaces.Services;
 using PlugHub.Shared.Models;
+using PlugHub.Shared.Models.Configuration;
 
 namespace PlugHub.UnitTests.Accessors.Configuration
 {
