+++ PlugHub.UnitTests/Services/Configuration/UserConfigServiceTests.cs
@@ -6,6 +6,7 @@ using PlugHub.Shared.Extensions;
 using PlugHub.Shared.Interfaces.Models;
 using PlugHub.Shared.Interfaces.Services;
 using PlugHub.Shared.Models;
+using PlugHub.Shared.Models.Configuration;
 
 namespace PlugHub.UnitTests.Services.Configuration
 {
